+++ packages/sls-core/.serverless/src/index.js
@@ -1,7 +1,7 @@
 const http = require('http')
 
 const server = http.createServer((req, res) => {
-  res.end('yyyyyyyyyyyyy99999999')
+  res.end('14:23')
 })
 
 server.listen(9000)
