+++ packages/sls-core/.serverless/src/test7.js
