+++ packages/sls-core/package.json
@@ -13,7 +13,7 @@
   "scripts": {
     "dev": "tsc --watch",
     "test": "tsc && node dist/index.js",
-    "test:scf": "cd .serverless && scf deploy --target scf --noCache",
+    "test:scf": "cd .serverless && scf deploy --target scf",
     "build": "rimraf lib && tsc"
   },
   "devDependencies": {
@@ -28,10 +28,12 @@
   "dependencies": {
     "@serverless/utils-china": "^1.1.7",
     "axios": "^1.3.5",
+    "dotenv": "^16.0.3",
     "globby": "^13.1.3",
     "graph-data-structure": "^3.3.0",
     "js-yaml": "^4.1.0",
     "jszip": "^3.10.1",
+    "p-limit": "^4.0.0",
     "serverless-cloud-framework": "^1.0.2",
     "tencentcloud-sdk-nodejs": "^4.0.575",
     "traverse": "^0.6.7",
