--- packages/sls-core/sls-src/index.js
@@ -1,7 +1,0 @@
-const http = require('http')
-
-const server = http.createServer((req, res) => {
-    res.end('xxxxx9999')
-})
-
-server.listen(9000)
