--- packages/sls-core/sls-src/scf_bootstrap
@@ -1,3 +1,0 @@
-#!/bin/sh
-/var/lang/node16/bin/node index.js
-
