+++ packages/sls-core/src/constants.ts
@@ -1,0 +1,3 @@
+export const FILE_SYLINK = 41471
+export const FILE_RW = 0o755
+export const FILE_R = 0o644
